Move all drawables to components folder
